refactor: 검색 결과 상태관리 hook 분리
